requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-crontab==0.7.1
djangorestframework==3.14.0
Fiona==1.9.3
geopandas==0.12.2
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
libretranslatepy==2.1.1
lxml==4.9.2
munch==2.5.0
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyOpenSSL==23.1.1
pyotp==2.8.0
pyproj==3.5.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
rfc3986==1.5.0
service-identity==21.1.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
translate==3.6.1
Twisted==22.10.0
txaio==23.1.1
typesense==0.15.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
zope.interface==6.0
